api_read/requirements.txt
mysql-connector-python==8.2.0
requests==2.31.0
Flask==3.0.0
Werkzeug==3.0.1
Flask-WTF==1.2.1
WTForms==3.1.1